Rebased and added attribute field to distinguish their characteristics for POST and DELETE
