+++ lib/pbench/client/__init__.py
@@ -38,6 +38,7 @@ class API(Enum):
     name completion.
     """
 
+    DATASETS = "datasets"
     DATASETS_CONTENTS = "datasets_contents"
     DATASETS_DATERANGE = "datasets_daterange"
     DATASETS_DELETE = "datasets_delete"
@@ -47,7 +48,6 @@ class API(Enum):
     DATASETS_MAPPINGS = "datasets_mappings"
     DATASETS_METADATA = "datasets_metadata"
     DATASETS_NAMESPACE = "datasets_namespace"
-    DATASETS_UPDATE = "datasets_update"
     DATASETS_SEARCH = "datasets_search"
     DATASETS_VALUES = "datasets_values"
     ENDPOINTS = "endpoints"
